LocalizationのExtensionをUtility内に統合
